[TASK-5] fix errors
