+++ import-service/serverless.yml
@@ -60,12 +60,12 @@ functions:
               querystrings:
                 name: true
           cors: true
-  importProductsFile:
-    handler: ./importfileparser/importfileparser.importFileParser
-    events:
-      - s3:
-          bucket: rs-task-5-uploaded
-          event: s3:ObjectCreated:*
+  # importFileParser:
+  #   handler: ./importfileparser/importfileparser.importFileParser
+  #   events:
+  #     - s3:
+  #         bucket: rs-task-5-uploaded
+  #         event: s3:ObjectCreated:*
 #    The following are a few example events you can configure
 #    NOTE: Please make sure to change your handler code to work with those events
 #    Check the event documentation for details
